+++ Common/src/main/java/com/afoxxvi/asteorbar/overlay/Overlays.java
@@ -3,8 +3,8 @@ package com.afoxxvi.asteorbar.overlay;
 import com.afoxxvi.asteorbar.AsteorBar;
 import com.afoxxvi.asteorbar.overlay.parts.*;
 import com.afoxxvi.asteorbar.utils.GuiHelper;
-import net.minecraft.client.Minecraft;
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.Minecraft;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -40,6 +40,12 @@ public class Overlays {
     public static boolean thirst = false;
     public static boolean mekanism = false;
     public static boolean dehydration = false;
+    public static boolean parcool = false;
+    public static boolean ironsSpellbooks = false;
+    public static boolean feathers = false;
+    public static boolean appleskin = false;
+    public static boolean superiorshields = false;
+    public static boolean vampirism = false;
 
     public static void init() {
         initialized = true;
@@ -47,6 +53,12 @@ public class Overlays {
         thirst = AsteorBar.platformAdapter.isModLoaded("thirst");
         mekanism = AsteorBar.platformAdapter.isModLoaded("mekanism");
         dehydration = AsteorBar.platformAdapter.isModLoaded("dehydration");
+        parcool = AsteorBar.platformAdapter.isModLoaded("parcool");
+        ironsSpellbooks = AsteorBar.platformAdapter.isModLoaded("irons_spellbooks");
+        feathers = AsteorBar.platformAdapter.isModLoaded("feathers");
+        appleskin = AsteorBar.platformAdapter.isModLoaded("appleskin");
+        superiorshields = AsteorBar.platformAdapter.isModLoaded("superiorshields");
+        vampirism = AsteorBar.platformAdapter.isModLoaded("vampirism");
     }
 
     public static void reset() {