+++ Fabric/src/main/java/com/afoxxvi/asteorbar/mixin/GuiMixin.java
@@ -1,11 +1,9 @@
 package com.afoxxvi.asteorbar.mixin;
 
 import com.afoxxvi.asteorbar.overlay.FabricGuiRegistry;
-import com.afoxxvi.asteorbar.overlay.FabricRenderGui;
 import com.afoxxvi.asteorbar.overlay.Overlays;
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.gui.Gui;
 import com.mojang.blaze3d.vertex.PoseStack;
+import net.minecraft.client.gui.Gui;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Invoker;
 import org.spongepowered.asm.mixin.injection.At;
@@ -27,20 +25,20 @@ public abstract class GuiMixin {
     }
 
     @Invoker("renderVehicleHealth")
-    public abstract void renderVehicleHealthRaw(PoseStack guiGraphics);
+    public abstract void renderVehicleHealthRaw(PoseStack poseStack);
 
     @Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;renderVehicleHealth(Lcom/mojang/blaze3d/vertex/PoseStack;)V"), method = "render")
-    public void renderVehicleHealth(Gui instance, PoseStack guiGraphics) {
+    public void renderVehicleHealth(Gui instance, PoseStack poseStack) {
         if (Overlays.style == Overlays.STYLE_NONE) {
-            renderVehicleHealthRaw(guiGraphics);
+            renderVehicleHealthRaw(poseStack);
         }
     }
 
 
     @Redirect(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;renderExperienceBar(Lcom/mojang/blaze3d/vertex/PoseStack;I)V"), method = "render")
-    public void renderExperienceBar(Gui instance, PoseStack guiGraphics, int i) {
+    public void renderExperienceBar(Gui instance, PoseStack poseStack, int i) {
         if (Overlays.style == Overlays.STYLE_NONE) {
-            instance.renderExperienceBar(guiGraphics, i);
+            instance.renderExperienceBar(poseStack, i);
         }
     }
 }